requirements.txt
orpheus-speech
colorama
whisper
omnisensevoice
openai
python-dotenv